Update README.md
Many people are getting hung up on the token handling portion of this package, not understanding that they need to setup an API endpoint. In my case, I wasn't sure what the API endpoint was actually supposed to do until I found some examples in the Stripe documentation. Including some quick links should cut down on the amount of "Issues" people put into the queue simply asking for help with this.